Move base unit tests to unit tests module
