+++ tests/unit/source_code/test_base.py
@@ -1,11 +1,12 @@
 import dataclasses
 
+from databricks.labs.ucx.hive_metastore.tables import Table
 from databricks.labs.ucx.source_code.base import (
     Advice,
     Advisory,
     Convention,
     Deprecation,
-    Failure,
+    Failure, UsedTable,
 )
 
 
@@ -40,3 +41,15 @@ def test_deprecation_initialization() -> None:
 def test_convention_initialization() -> None:
     convention = Convention('code5', 'This is a convention', 1, 1, 2, 2)
     assert isinstance(convention, Advice)
+
+
+def test_used_table_from_table() -> None:
+    table = Table("catalog", "schema", "table", "MANAGED", "DELTA")
+
+    used_table = UsedTable.from_table(table, is_read=False, is_write=True)
+
+    assert used_table.catalog_name == "catalog"
+    assert used_table.schema_name == "schema"
+    assert used_table.table_name == "table"
+    assert not used_table.is_read
+    assert used_table.is_write
